requirements.txt
Django==2.2.7
djangorestframework==3.10.3
django-oauth-toolkit==1.2.0
mysqlclient==1.4.5
PyYAML==5.1.2
